% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minidlyr-package.R
\docType{package}
\name{minidlyr-package}
\alias{minidlyr}
\alias{minidlyr-package}
\title{minidlyr: Reimplementing Dplyr (or not)}
\description{
Useless thing.
}
\author{
\strong{Maintainer}: Me Me \email{me.me@example.com}

}
\keyword{internal}
