% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{my_fancy_add}
\alias{my_fancy_add}
\title{Title}
\usage{
my_fancy_add(x, y)
}
\arguments{
\item{x}{Is first variable}

\item{y}{Is second variable}
}
\value{
The sum of x anf y
}
\description{
This will be the description.
}
\details{
This will be the details.
}
\examples{
my_fancy_add(1, 2)
my_fancy_add(1:2, 2:3)
}
