% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select2.R
\name{select2}
\alias{select2}
\title{Select variable}
\usage{
select2(df, ind)
}
\arguments{
\item{df}{A data frame.}

\item{ind}{A vector of character, integer, boolean}
}
\value{
A data frame.
}
\description{
A poor version of the select() from dplyr.
This fonction allow to select column in a data.frame.
}
\examples{
select2(cars, "speed")

select2(cars, 1:2)
}
